+++ .changeset/angry-boats-kick.md
@@ -1,0 +1,5 @@
+---
+"@saleor/app-sdk": major
+---
+
+Changed publically exported paths. New exports will be documented in docs and migration guide
