+++ package.json
@@ -143,11 +143,6 @@
       "import": "./settings-manager/index.mjs",
       "require": "./settings-manager/index.js"
     },
-    "./urls": {
-      "types": "./urls.d.ts",
-      "import": "./urls.mjs",
-      "require": "./urls.js"
-    },
     "./app-bridge": {
       "types": "./app-bridge/index.d.ts",
       "import": "./app-bridge/index.mjs",
@@ -188,15 +183,15 @@
       "import": "./saleor-app.mjs",
       "require": "./saleor-app.js"
     },
-    "./verify-jwt": {
-      "types": "./verify-jwt.d.ts",
-      "import": "./verify-jwt.mjs",
-      "require": "./verify-jwt.js"
+    "./auth/node": {
+      "types": "./auth/node.index.d.ts",
+      "import": "./auth/node/index.mjs",
+      "require": "./auth/node/index.js"
     },
-    "./verify-signature": {
-      "types": "./verify-signature.d.ts",
-      "import": "./verify-signature.mjs",
-      "require": "./verify-signature.js"
+    "./auth/browser": {
+      "types": "./auth/node.index.d.ts",
+      "import": "./auth/node/index.mjs",
+      "require": "./auth/node/index.js"
     },
     "./headers": {
       "types": "./headers.d.ts",
@@ -207,6 +202,11 @@
       "types": "./util/public/index.d.ts",
       "import": "./util/public/index.mjs",
       "require": "./util/public/index.js"
+    },
+    "./util/browser": {
+      "types": "./util/public/browser/index.d.ts",
+      "import": "./util/public/browser/index.mjs",
+      "require": "./util/public/browser/index.js"
     },
     "./types": {
       "types": "./types.d.ts"
