+++ src/auth/browser/verify-jwt.ts
@@ -1,8 +1,9 @@
 import * as jose from "jose";
 
+import { getJwksUrlFromSaleorApiUrl } from "@/auth/node";
+
 import { createDebug } from "../../debug";
 import { Permission } from "../../types";
-import { getJwksUrlFromSaleorApiUrl } from "../../urls";
 import { hasPermissionsInJwtToken } from "./has-permissions-in-jwt-token";
 import { verifyTokenExpiration } from "./verify-token-expiration";
