+++ src/auth/node/verify-signature.ts
@@ -1,44 +1,8 @@
 import * as jose from "jose";
 
-import { createDebug } from "./debug";
-import { getJwksUrlFromSaleorApiUrl } from "./urls";
+import { createDebug } from "../../debug";
 
 const debug = createDebug("verify-signature");
-
-/**
- * Verify Webhook payload signature with public key of given `domain`
- * https://docs.saleor.io/docs/3.x/developer/extending/apps/asynchronous-webhooks#payload-signature
- *
- * Use Saleor URL to fetch JWKS
- *
- * TODO: Add test
- */
-export const verifySignatureFromApiUrl = async (
-  saleorApiUrl: string,
-  signature: string,
-  rawBody: string
-) => {
-  const [header, , jwsSignature] = signature.split(".");
-  const jws: jose.FlattenedJWSInput = {
-    protected: header,
-    payload: rawBody,
-    signature: jwsSignature,
-  };
-
-  const remoteJwks = jose.createRemoteJWKSet(
-    new URL(getJwksUrlFromSaleorApiUrl(saleorApiUrl))
-  ) as jose.FlattenedVerifyGetKey;
-
-  debug("Created remote JWKS");
-
-  try {
-    await jose.flattenedVerify(jws, remoteJwks);
-    debug("JWKS verified");
-  } catch {
-    debug("JWKS verification failed");
-    throw new Error("JWKS verification failed");
-  }
-};
 
 /**
  * Verify the Webhook payload signature from provided JWKS string.
@@ -74,3 +38,6 @@ export const verifySignatureWithJwks = async (jwks: string, signature: string, r
     throw new Error("JWKS verification failed");
   }
 };
+
+export const getJwksUrlFromSaleorApiUrl = (saleorApiUrl: string): string =>
+  `${new URL(saleorApiUrl).origin}/.well-known/jwks.json`;
