--- src/const.ts
@@ -1,7 +1,0 @@
-export const SALEOR_EVENT_HEADER = "saleor-event";
-export const SALEOR_SIGNATURE_HEADER = "saleor-signature";
-export const SALEOR_AUTHORIZATION_BEARER_HEADER = "authorization-bearer";
-export const SALEOR_API_URL_HEADER = "saleor-api-url";
-export const SALEOR_SCHEMA_VERSION = "saleor-schema-version";
-
-export * from "./locales";
