+++ src/handlers/platforms/fetch-api/create-app-register-handler.test.ts
@@ -1,9 +1,9 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { AuthData } from "@/APL";
-import { SALEOR_API_URL_HEADER } from "@/const";
-import * as fetchRemoteJwksModule from "@/fetch-remote-jwks";
+import * as fetchRemoteJwksModule from "@/auth/node/fetch-remote-jwks";
 import * as getAppIdModule from "@/get-app-id";
+import { SALEOR_API_URL_HEADER } from "@/headers";
 import { MockAPL } from "@/test-utils/mock-apl";
 
 import {
@@ -114,19 +114,19 @@ describe("Fetch API createAppRegisterHandler", () => {
         expect.objectContaining({
           authToken,
           saleorApiUrl,
-        })
+        }),
       );
       expect(mockOnRequestVerified).toHaveBeenCalledWith(
         request,
         expect.objectContaining({
           authData: expectedAuthData,
-        })
+        }),
       );
       expect(mockOnAuthAplSaved).toHaveBeenCalledWith(
         request,
         expect.objectContaining({
           authData: expectedAuthData,
-        })
+        }),
       );
       expect(mockOnAuthAplFailed).not.toHaveBeenCalled();
     });
@@ -150,7 +150,7 @@ describe("Fetch API createAppRegisterHandler", () => {
         expect.objectContaining({
           error: expect.any(Error),
           authData: expectedAuthData,
-        })
+        }),
       );
     });
 
@@ -161,7 +161,7 @@ describe("Fetch API createAppRegisterHandler", () => {
           context.respondWithError({
             status: 401,
             message: "test message",
-          })
+          }),
         );
       const handler = createAppRegisterHandler({
         apl: mockApl,
