+++ src/handlers/platforms/next/create-manifest-handler.test.ts
@@ -1,7 +1,7 @@
 import { createMocks } from "node-mocks-http";
 import { describe, expect, it, vi } from "vitest";
 
-import { SALEOR_SCHEMA_VERSION } from "@/const";
+import { SALEOR_SCHEMA_VERSION } from "@/headers";
 
 import { createManifestHandler, CreateManifestHandlerOptions } from "./create-manifest-handler";
