+++ src/handlers/shared/saleor-request-processor.test.ts
@@ -6,7 +6,7 @@ import {
   SALEOR_EVENT_HEADER,
   SALEOR_SCHEMA_VERSION,
   SALEOR_SIGNATURE_HEADER,
-} from "@/const";
+} from "@/headers";
 import { MockAdapter } from "@/test-utils/mock-adapter";
 
 import { SaleorRequestProcessor } from "./saleor-request-processor";
