+++ src/headers.ts
@@ -1,10 +1,8 @@
-import {
-  SALEOR_API_URL_HEADER,
-  SALEOR_AUTHORIZATION_BEARER_HEADER,
-  SALEOR_EVENT_HEADER,
-  SALEOR_SCHEMA_VERSION,
-  SALEOR_SIGNATURE_HEADER,
-} from "./const";
+export const SALEOR_EVENT_HEADER = "saleor-event";
+export const SALEOR_SIGNATURE_HEADER = "saleor-signature";
+export const SALEOR_AUTHORIZATION_BEARER_HEADER = "authorization-bearer";
+export const SALEOR_API_URL_HEADER = "saleor-api-url";
+export const SALEOR_SCHEMA_VERSION = "saleor-schema-version";
 
 const toStringOrUndefined = (value: string | string[] | undefined) =>
   value ? value.toString() : undefined;
