--- src/urls.ts
@@ -1,2 +1,0 @@
-export const getJwksUrlFromSaleorApiUrl = (saleorApiUrl: string): string =>
-  `${new URL(saleorApiUrl).origin}/.well-known/jwks.json`;
