--- src/util/index.ts
@@ -1,3 +1,0 @@
-export * from "./is-in-iframe";
-export * from "./schema-version";
-export * from "./use-is-mounted";
