+++ tsup.config.ts
@@ -2,14 +2,11 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   entry: {
-    const: "src/const.ts",
     types: "src/types.ts",
-    urls: "src/urls.ts",
     headers: "src/headers.ts",
     util: "src/util/public/index.ts",
+    "util/browser": "src/util/public/browser/index.ts",
     "saleor-app": "src/saleor-app.ts",
-    "verify-jwt": "src/verify-jwt.ts",
-    "verify-signature": "src/verify-signature.ts",
     /**
      * APLs
      */
@@ -31,6 +28,8 @@ export default defineConfig({
     "handlers/fetch-api/index": "src/handlers/platforms/fetch-api/index.ts",
     "handlers/aws-lambda/index": "src/handlers/platforms/aws-lambda/index.ts",
     "handlers/next-app-router/index": "src/handlers/platforms/next-app-router/index.ts",
+    "auth/node/index": "src/auth/node/index.ts",
+    "auth/browser/index": "src/auth/browser/index.ts",
   },
   dts: true,
   clean: true,
